// src/API/AuthApi/auth-api-routes.enum.ts
export enum AuthApiRoutes {
  auth = '/auth',
  emailExist = '/auth/emailExist',
  loginRequest = '/auth/loginRequest',
  checkDetails = '/auth/checkDetails',
  createAccount = '/auth/createAccount',
  sendTempCode = '/auth/sendTempCode',
  controlQuestion = '/auth/controlQuestions',
  // resetVisitorData = '/auth/reset-visitor-data',
  createAccountData = '/auth/createAccountData',
}